refactor: refactor productContext and add product route with param
