+++ src/component/productPage/index.tsx
@@ -5,32 +5,42 @@ import Button from '@mui/material/Button';
 import Grid from '@mui/material/Grid';
 import Divider from '@mui/material/Divider';
 import Typography from '@mui/material/Typography';
-import Picture from '../../images/goods.jpeg';
+import {useParams} from "react-router-dom";
+import {useContext} from "react";
+import {ProductContext} from "../../App";
+
+export default function ProductPage(this: any) {
+    const products = useContext(ProductContext);
+    let param = useParams();
+    const product = products.filter((item) => {
+        return item.id === Object.values(param)[0];
+    })
 
 
-export default function ProductPage() {
     return (
         <Container maxWidth={"lg"}>
             <Box sx={{ width: '100%', bgcolor: 'background.paper' }}>
                 <Box sx={{ m: 2 }}>
-                    <img src={Picture} alt="图片"/>
+                    <img src={require("../../images/" + product[0].img + '.jpeg')} alt="图片"/>
                 </Box>
                 <Box sx={{ my: 3, mx: 2 }}>
                     <Grid container alignItems="center">
                         <Grid item xs>
                             <Typography gutterBottom variant="h4" component="div">
-                                Toothbrush
+                                {/*@ts-ignore*/}
+                                {product[0].name}
                             </Typography>
                         </Grid>
                         <Grid item>
                             <Typography gutterBottom variant="h6" component="div">
-                                $4.50
+                                {/*@ts-ignore*/}
+                                ¥{product[0].price}
                             </Typography>
                         </Grid>
                     </Grid>
                     <Typography color="text.secondary" variant="body2">
-                        Pinstriped cornflower blue cotton blouse takes you on a walk to the park or
-                        just down the hall.
+                        {/*@ts-ignore*/}
+                        {product[0].description}
                     </Typography>
                 </Box>
                 <Divider variant="middle" />
